% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{theme_comidistar}
\alias{theme_comidistar}
\title{Tema de ggplot2 de comidistar}
\usage{
theme_comidistar()
}
\description{
Tema de ggplot2 de comidistar
}
\examples{
library(ggplot2)
ggplot(puntuaciones, aes(puntuacion, producto)) + geom_point() + theme_comidistar()
}
\author{
Gonzalo Garcia-Castro
}
