% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/videos.R
\docType{data}
\name{videos}
\alias{videos}
\title{Fecha y URL de los vídeos de las Catas de Expertos de El Comidista}
\format{
Un tibble con 38 filas y 3 columnas:
\itemize{
\item producto: producto puntuado
\item fecha: fecha de subida del vídeo a YouTube
\item link: URL del vídeo
}
}
\source{
\url{https://elcomidista.elpais.com/agr/catas_de_expertos/a}
}
\usage{
videos
}
\description{
Incluye la fecha y URL de los vídeos de los que fueron extraídos los datos que
contiene este paquete.
}
\author{
Gonzalo Garcia-Castro
}
\keyword{datasets}
