% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comidistar-package.R
\docType{package}
\name{comidistar-package}
\alias{comidistar}
\alias{comidistar-package}
\title{comidistar: Base de datos con las puntuaciones de las catas a ciegas de El Comidista}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

comidistar es un paquete de R que contiene una base de datos con las puntuaciones de las catas a ciegas de El Comidista. Las puntuaciones y datos adicionales sobre los productos de las catas fueron transcritos manualmente a partir de los vídeos subidos al canal de YouTube de El Comidista.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gongcastro.github.io/comidistar/}
  \item \url{https://github.com/gongcastro/comidistar/}
  \item Report bugs at \url{https://github.com/gongcastro/comidistar/issues}
}

}
\author{
\strong{Maintainer}: Gonzalo Garcia-Castro \email{gonzaloggc95@gmail.com} (\href{https://orcid.org/0000-0002-8553-4209}{ORCID})

}
\keyword{internal}
