% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comidistar-package.R
\docType{package}
\name{comidistar-package}
\alias{comidistar}
\alias{comidistar-package}
\title{comidistar: Puntuaciones de las Catas a Ciegas de El Comidista}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Base de datos documentada con las puntuaciones de las catas a ciegas de El Comidista, información sobre los productos y su precio, y metadatos sobre el vídeo subido a YouTube.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ropensci.github.io/comidistar/}
  \item \url{https://github.com/ropensci/comidistar/}
  \item Report bugs at \url{https://github.com/ropensci/comidistar/issues}
}

}
\author{
\strong{Maintainer}: Gonzalo Garcia-Castro \email{gongarciacastro@gmail.com} (\href{https://orcid.org/0000-0002-8553-4209}{ORCID})

}
\keyword{internal}
