% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comidistar-package.R
\docType{package}
\name{comidistar-package}
\alias{comidistar}
\alias{comidistar-package}
\title{comidistar: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
